Added genai table. Modified homescreen toasts. (#23367)
Preview isn't working right now as my personal repo doesn't want to run
actions - please make a codespace to test it!

Should suffice as a genai install table, and added nifty things like a
'copy' function :)
