+++ src/routes/components/hero.svelte
@@ -6,6 +6,7 @@
 	import { Highlight } from 'svelte-highlight';
 	import { bash } from 'svelte-highlight/languages';
 	import FaRegClipboard from 'svelte-icons/fa/FaRegClipboard.svelte';
+	import FaClipboardCheck from 'svelte-icons/fa/FaClipboardCheck.svelte'
 	import OnnxLight from '../../images/ONNX-Light.svelte';
 	import OnnxDark from '../../images/ONNX-Dark.svelte';
 	import { fade } from 'svelte/transition';
@@ -55,24 +56,9 @@
 {#if copied}
 	<div class="toast toast-top top-14 z-50" role="alert">
 		<div class="alert alert-info">
-			<svg
-				xmlns="https://www.w3.org/2000/svg"
-				width="16"
-				height="16"
-				fill="currentColor"
-				class="bi bi-clipboard2-check"
-				viewBox="0 0 16 16"
-			>
-				<path
-					d="M9.5 0a.5.5 0 0 1 .5.5.5.5 0 0 0 .5.5.5.5 0 0 1 .5.5V2a.5.5 0 0 1-.5.5h-5A.5.5 0 0 1 5 2v-.5a.5.5 0 0 1 .5-.5.5.5 0 0 0 .5-.5.5.5 0 0 1 .5-.5h3Z"
-				/>
-				<path
-					d="M3 2.5a.5.5 0 0 1 .5-.5H4a.5.5 0 0 0 0-1h-.5A1.5 1.5 0 0 0 2 2.5v12A1.5 1.5 0 0 0 3.5 16h9a1.5 1.5 0 0 0 1.5-1.5v-12A1.5 1.5 0 0 0 12.5 1H12a.5.5 0 0 0 0 1h.5a.5.5 0 0 1 .5.5v12a.5.5 0 0 1-.5.5h-9a.5.5 0 0 1-.5-.5v-12Z"
-				/>
-				<path
-					d="M10.854 7.854a.5.5 0 0 0-.708-.708L7.5 9.793 6.354 8.646a.5.5 0 1 0-.708.708l1.5 1.5a.5.5 0 0 0 .708 0l3-3Z"
-				/>
-			</svg>
+			<div class="icon" style="width: 16px; height: 16px;">
+				<FaClipboardCheck />
+			</div>
 			<span>Code successfully copied!</span>
 		</div>
 	</div>
